requirements.txt
streamlit
langchain
langgraph
groq
python-dotenv
langchain-openai
faiss-cpu
langchain_community
langchain_groq
sentence-transformers